Tidy up escape handling
