+++ osu.Game/Graphics/UserInterface/OsuTextBox.cs
@@ -9,10 +9,12 @@ using osu.Framework.Input;
 using osu.Game.Graphics.Sprites;
 using OpenTK.Graphics;
 using osu.Framework.Extensions.Color4Extensions;
+using osu.Framework.Input.Bindings;
+using osu.Game.Input.Bindings;
 
 namespace osu.Game.Graphics.UserInterface
 {
-    public class OsuTextBox : TextBox
+    public class OsuTextBox : TextBox, IKeyBindingHandler<GlobalAction>
     {
         protected override Color4 BackgroundUnfocused => Color4.Black.Opacity(0.5f);
         protected override Color4 BackgroundFocused => OsuColour.Gray(0.3f).Opacity(0.8f);
@@ -33,10 +35,7 @@ namespace osu.Game.Graphics.UserInterface
             TextContainer.Height = 0.5f;
             CornerRadius = 5;
 
-            Current.DisabledChanged += disabled =>
-            {
-                Alpha = disabled ? 0.3f : 1;
-            };
+            Current.DisabledChanged += disabled => { Alpha = disabled ? 0.3f : 1; };
         }
 
         [BackgroundDependencyLoader]
@@ -59,5 +58,21 @@ namespace osu.Game.Graphics.UserInterface
         }
 
         protected override Drawable GetDrawableCharacter(char c) => new OsuSpriteText { Text = c.ToString(), TextSize = CalculatedTextSize };
+
+        public bool OnPressed(GlobalAction action)
+        {
+            if (action == GlobalAction.Back)
+            {
+                if (Text.Length > 0)
+                    Text = string.Empty;
+                else
+                    KillFocus();
+                return true;
+            }
+
+            return true;
+        }
+
+        public bool OnReleased(GlobalAction action) => false;
     }
 }
