+++ osu.Game/Screens/OsuScreen.cs
@@ -8,7 +8,6 @@ using osu.Framework.Audio;
 using osu.Framework.Audio.Sample;
 using osu.Framework.Configuration;
 using osu.Framework.Graphics;
-using osu.Framework.Input;
 using osu.Framework.Input.Bindings;
 using osu.Framework.Screens;
 using osu.Game.Beatmaps;
@@ -17,7 +16,6 @@ using osu.Game.Input.Bindings;
 using osu.Game.Rulesets;
 using osu.Game.Screens.Menu;
 using OpenTK;
-using OpenTK.Input;
 using osu.Game.Overlays;
 using osu.Framework.Graphics.Containers;
 
@@ -106,6 +104,8 @@ namespace osu.Game.Screens
 
         public bool OnPressed(GlobalAction action)
         {
+            if (!IsCurrentScreen) return false;
+
             if (action == GlobalAction.Back && AllowBackButton)
             {
                 Exit();
@@ -116,20 +116,6 @@ namespace osu.Game.Screens
         }
 
         public bool OnReleased(GlobalAction action) => action == GlobalAction.Back && AllowBackButton;
-
-        protected override bool OnKeyDown(InputState state, KeyDownEventArgs args)
-        {
-            if (args.Repeat || !IsCurrentScreen) return false;
-
-            switch (args.Key)
-            {
-                case Key.Escape:
-                    Exit();
-                    return true;
-            }
-
-            return base.OnKeyDown(state, args);
-        }
 
         protected override void OnResuming(Screen last)
         {
